tests: support blockchain_tests
